[network] Move UpdatableIDProvider to tests
