+++ network/test/updatable_provider.go
@@ -1,4 +1,4 @@
-package id
+package test
 
 import (
 	"sync"
